implement update customer method

+++ src/main/java/org/example/repository/CustomerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<CustomerEntity, Long> {
+public interface CustomerRepository extends CrudRepository<CustomerEntity, Integer> {
 
 }
